Remove tedge_config_repository.rs
All its content was related to TEdgeConfigLocation.
