+++ crates/common/tedge_config/src/tedge_config_cli/mod.rs
@@ -1,7 +1,6 @@
 pub mod config_setting;
 pub mod error;
 pub mod tedge_config_location;
-pub mod tedge_config_repository;
 
 mod figment;
 pub mod models;
